pyproject.toml
[tool.poetry]
name = "moo_classification"
version = "0.22.0"
description = ""
authors = ["Eike Mentzendorff <eike.mentzendorff@spryfox.de>", "Sarosh Manzoor <sarosh.manzoor@spryfox.de>"]
readme = "README.md"
packages = [{ include = "moo_classification" }]

[tool.poetry.dependencies]
python = "3.10.8"
torch = "^2.3.1"
pandas = "1.5.3"
numpy = "1.23.4"
tqdm = "4.64.1"
pyyaml = "^6.0.1"
neptune = "1.10.4"
matplotlib = "^3.9.1"
torchvision = "^0.18.1"
line-profiler = "^4.1.3"
seaborn = "^0.13.2"
plotly = "5.22.0"
pytest = "^8.2.2"
lightgbm = "^4.4.0"
scikit-learn = "1.5.1"
pydantic = "^2.8.2"
treeboost-autograd = "^0.1.2"
toml = "^0.10.2"
vaex = { version = "4.17.0", allow-prereleases = true }

[tool.poetry.group.dev.dependencies]
setuptools = "^70.1.0"
black = { extras = ["d"], version = "^24.4.2" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
